Fix new scheduling route flow
