+++ platform/api-server/api/routes.js
@@ -45,7 +45,7 @@ module.exports = function(app) {
     app.get('/platform/link-graph-visual', linkGraphController.renderLinkGraph);
     app.post('/platform/schedule-app',
              uploader.fields([{ name: 'deployMetadata' }, { name: 'appPackage' }]),
-             platformAPIController.deployApplication);
+             platformAPIController.scheduleApplication);
     app.post('/platform/disseminate-all', platformAPIController.disseminateAll);
     app.post('/platform/query-all', platformAPIController.queryAll);
     app.post('/platform/update-privacy-policy',
